Removed deadname
What a shocking turn of events! Who could have possibly foreseen this! Apparently not me up until a few months ago. Anyhow, hi whoever is reading this, I doubt it will be me. Maybe it will be! Always interesting finding messages like this. I need to be working on a game for uni (a racing game for GDS1 lol), but here I am updating a bunch of websites and stuff. Very funky.